Move api/v1 package to internal/types
api/v1 is to be reserved for HTTP API types. This commit moves the
internal representation of issuers to an internal package so the
exported interface represents what client libraries can use.
