+++ internal/jwks/strategy.go
@@ -5,8 +5,8 @@ import (
 	"context"
 	"fmt"
 
-	v1 "go.infratographer.com/identity-manager-sts/pkg/api/v1"
 	"go.infratographer.com/identity-manager-sts/internal/fositex"
+	"go.infratographer.com/identity-manager-sts/internal/types"
 )
 
 var (
@@ -15,11 +15,11 @@ var (
 )
 
 type issuerJWKSURIStrategy struct {
-	issuerSvc v1.IssuerService
+	issuerSvc types.IssuerService
 }
 
 // NewIssuerJWKSURIStrategy creates a new fosite.IssuerJWKSURIStrategy.
-func NewIssuerJWKSURIStrategy(issuerSvc v1.IssuerService) fositex.IssuerJWKSURIStrategy {
+func NewIssuerJWKSURIStrategy(issuerSvc types.IssuerService) fositex.IssuerJWKSURIStrategy {
 	out := issuerJWKSURIStrategy{
 		issuerSvc: issuerSvc,
 	}
