+++ internal/types/types.go
@@ -1,4 +1,4 @@
-package v1
+package types
 
 import (
 	"context"
@@ -25,6 +25,7 @@ type Issuer struct {
 
 // IssuerService represents a service for managing issuers.
 type IssuerService interface {
+	Create(ctx context.Context, iss *Issuer) (*Issuer, error)
 	GetByURI(ctx context.Context, uri string) (*Issuer, error)
 }
